add golike defer and some xen api improvements with structuration
